Use GHC to compile fnv.c to avoid link errors
GHC 7.8.3 on Windows has problems to link to MinGW-generated object files,
depending on the version of GCC. As GHC can compile C files (it bundles
the correct version of MinGW with Haskell Platform), let's use it to
compile the fnv.c file. And let's do it for all platforms.

This is related to http://ghc.haskell.org/trac/ghc/ticket/9080